refactor: mega file is gone, now we have some little files
